Move Template class into Audits module
This is related to the CSV export of reporting. Again this may need
to move into domain alongside report class, but we are making this
initial change for now and will revisit in the future.
